More sites supported
generally a trend we've seen is sites commonly have with and height of 100% for their embedded video so this is the first change that makes this friendlier to more sites. I assume at some point we'll find a method to get embeds that is faster and more efficient but for the time being we're sticking with this.